Override start method in Authenticator
This commit patchs the start method in the LoginFormAuthenticator.

Without this patch, we can observe that ApiPlatform/Symfony tries to redirect to the home page instead of throwing an Unauthorized Error to the user.

This patch now throws an error when symfony tries to redirect to an API Route which is protected